refactor: moved experiences into an async component with loading and error catching
